pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gw-mos"
version = "0.1.0"
description = "CLI-first research paper completion workflow for theory-led manuscripts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7",
  "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]

[project.scripts]
gw-mos = "gw_mos.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
